Refactored the roundrobin scheduler into a dedicated sub crate.

+++ main64/kernel_rust/tests/syscall_dispatch_test.rs
@@ -528,7 +528,7 @@ fn test_write_serial_rejects_len_crossing_canonical_boundary() {
     // ptr is near the top of user space; ptr + 8193 crosses USER_CANONICAL_END.
     // After clamping to 4096, the access would be safe — but we must reject the
     // full claimed range before applying the DoS cap.
-    let ptr = (0x0000_8000_0000_0000u64 - 4096) as u64; // last 4 KiB of user space
+    let ptr = 0x0000_8000_0000_0000u64 - 4096; // last 4 KiB of user space
     let ret = syscall::dispatch(
         SyscallId::WriteSerial as u64,
         ptr,
@@ -561,7 +561,7 @@ fn test_write_console_rejects_overflowing_len() {
 /// Contract: write_console rejects a (ptr, len) pair that crosses the user canonical boundary.
 #[test_case]
 fn test_write_console_rejects_len_crossing_canonical_boundary() {
-    let ptr = (0x0000_8000_0000_0000u64 - 4096) as u64;
+    let ptr = 0x0000_8000_0000_0000u64 - 4096;
     let ret = syscall::dispatch(
         SyscallId::WriteConsole as u64,
         ptr,
